+++ android/src/main/kotlin/com/mirrorfly/mirrorfly_plugin/FlyMethodConstants.kt
@@ -284,7 +284,8 @@ object FlyMethodConstants {
         "restoreBackup" to flyChatMethods::restoreBackup,
         "cancelBackup" to flyChatMethods::cancelBackup,
         "cancelRestore" to flyChatMethods::cancelRestore,
-    )
+        "setTranslations" to flyChatMethods::setTranslations,
+        )
 
     private val callEventListeners: Map<String, EventChannel.StreamHandler> = mapOf(
         Constants.onLocalVideoTrackAdded to EventStreamHandler(),