+++ ios/Classes/FlyEventChannelInitializer.swift
@@ -141,7 +141,6 @@ class FlyChatEventChannelInitializer {
         (channelName: Constants.onSuperAdminDeleteGroupChannel, streamHandler: OnSuperAdminDeleteGroupChannelStreamHandler()),
     ]
     var sinkValues: [String: String] = [:]
-    
     func initializeChatEventChannels(registrar: FlutterPluginRegistrar) {
         for (channelName, streamHandler) in FlyChatEventChannelInitializer.chatEventChannels {
                 let handler: (NSObjectProtocol & FlutterStreamHandler)? = streamHandler